+++ mockall/tests/automock_nondebug.rs
@@ -1,0 +1,27 @@
+// vim: tw=80
+//! A method may have non-Debug arguments and/or return values.
+#![deny(warnings)]
+
+use mockall::*;
+
+// Don't derive Debug
+pub struct NonDebug(u32);
+
+#[automock]
+pub trait Foo {
+    fn foo(&self, x: NonDebug);
+}
+
+#[test]
+#[cfg_attr(feature = "nightly", should_panic(
+        expected = "MockFoo::foo(?): No matching expectation found"
+))]
+#[cfg_attr(not(feature = "nightly"), should_panic(
+        expected = "MockFoo::foo(?): No matching expectation found"
+))]
+fn with_no_matches() {
+    let mock = MockFoo::new();
+    mock.foo(NonDebug(5));
+}
+
+