+++ mockall/tests/mock_struct.rs
@@ -64,7 +64,12 @@ mod checkpoint {
     }
 
     #[test]
-    #[should_panic(expected = "MockFoo::foo: No matching expectation found")]
+    #[cfg_attr(feature = "nightly", should_panic(
+            expected = "MockFoo::foo(0): No matching expectation found"
+    ))]
+    #[cfg_attr(not(feature = "nightly"), should_panic(
+            expected = "MockFoo::foo(?): No matching expectation found"
+    ))]
     fn removes_old_expectations() {
         let mut mock = MockFoo::new();
         mock.expect_foo()
@@ -118,7 +123,12 @@ mod r#match {
     }
 
     #[test]
-    #[should_panic(expected = "MockFoo::bar: No matching expectation found")]
+    #[cfg_attr(feature = "nightly", should_panic(
+            expected = "MockFoo::bar(5): No matching expectation found"
+    ))]
+    #[cfg_attr(not(feature = "nightly"), should_panic(
+            expected = "MockFoo::bar(?): No matching expectation found"
+    ))]
     fn with_no_matches() {
         let mut mock = MockFoo::new();
         mock.expect_bar()
@@ -146,7 +156,12 @@ mod r#match {
     }
 
     #[test]
-    #[should_panic(expected = "MockFoo::bar: No matching expectation found")]
+    #[cfg_attr(feature = "nightly", should_panic(
+            expected = "MockFoo::bar(5): No matching expectation found"
+    ))]
+    #[cfg_attr(not(feature = "nightly"), should_panic(
+            expected = "MockFoo::bar(?): No matching expectation found"
+    ))]
     fn withf_no_matches() {
         let mut mock = MockFoo::new();
         mock.expect_bar()